+++ hc/front/tests/test_add_sms.py
@@ -1,5 +1,5 @@
 from django.test.utils import override_settings
-from hc.api.models import Channel
+from hc.api.models import Channel, Check
 from hc.test import BaseTestCase
 
 
@@ -7,11 +7,13 @@ from hc.test import BaseTestCase
 class AddSmsTestCase(BaseTestCase):
     def setUp(self):
         super().setUp()
-        self.url = "/projects/%s/add_sms/" % self.project.code
+        self.check = Check.objects.create(project=self.project)
+        self.url = f"/projects/{self.project.code}/add_sms/"
 
     def test_instructions_work(self):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.get(self.url)
+        self.assertContains(r, "Add SMS Integration")
         self.assertContains(r, "Get a SMS message")
 
     @override_settings(USE_PAYMENTS=True)
@@ -38,15 +40,18 @@ class AddSmsTestCase(BaseTestCase):
         self.assertFalse(c.sms_notify_up)
         self.assertEqual(c.project, self.project)
 
+        # Make sure it calls assign_all_checks
+        self.assertEqual(c.checks.count(), 1)
+
     def test_it_rejects_bad_number(self):
-        for v in ["not a phone number address", False, 15, "+123456789A"]:
+        for v in ["not a phone number", False, 15, "+123456789A"]:
             form = {"phone": v}
             self.client.login(username="alice@example.org", password="password")
             r = self.client.post(self.url, form)
             self.assertContains(r, "Invalid phone number format.")
 
     def test_it_trims_whitespace(self):
-        form = {"phone": "   +1234567890   "}
+        form = {"phone": "   +1234567890   ", "down": True}
 
         self.client.login(username="alice@example.org", password="password")
         self.client.post(self.url, form)
@@ -69,7 +74,7 @@ class AddSmsTestCase(BaseTestCase):
         self.assertEqual(r.status_code, 403)
 
     def test_it_strips_invisible_formatting_characters(self):
-        form = {"label": "My Phone", "phone": "\u202c+1234567890\u202c"}
+        form = {"phone": "\u202c+1234567890\u202c", "down": True}
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, form)
@@ -79,7 +84,7 @@ class AddSmsTestCase(BaseTestCase):
         self.assertEqual(c.phone_number, "+1234567890")
 
     def test_it_strips_hyphens(self):
-        form = {"label": "My Phone", "phone": "+123-4567890"}
+        form = {"phone": "+123-4567890", "down": True}
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, form)
@@ -89,7 +94,7 @@ class AddSmsTestCase(BaseTestCase):
         self.assertEqual(c.phone_number, "+1234567890")
 
     def test_it_strips_spaces(self):
-        form = {"label": "My Phone", "phone": "+123 45 678 90"}
+        form = {"phone": "+123 45 678 90", "down": True}
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, form)
@@ -98,16 +103,19 @@ class AddSmsTestCase(BaseTestCase):
         c = Channel.objects.get()
         self.assertEqual(c.phone_number, "+1234567890")
 
-    def test_it_obeys_up_down_flags(self):
-        form = {"label": "My Phone", "phone": "+1234567890"}
+    def test_it_handles_down_false_up_true(self):
+        form = {"phone": "+1234567890", "up": True}
 
         self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, form)
-        self.assertRedirects(r, self.channels_url)
+        self.client.post(self.url, form)
 
         c = Channel.objects.get()
-        self.assertEqual(c.kind, "sms")
-        self.assertEqual(c.phone_number, "+1234567890")
-        self.assertEqual(c.name, "My Phone")
         self.assertFalse(c.sms_notify_down)
-        self.assertFalse(c.sms_notify_up)
+        self.assertTrue(c.sms_notify_up)
+
+    def test_it_rejects_unchecked_up_and_down(self):
+        form = {"phone": "+1234567890"}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, form)
+        self.assertContains(r, "Please select at least one.")