+++ hc/front/tests/test_add_whatsapp.py
@@ -1,5 +1,5 @@
 from django.test.utils import override_settings
-from hc.api.models import Channel
+from hc.api.models import Channel, Check
 from hc.test import BaseTestCase
 
 TEST_CREDENTIALS = {
@@ -14,11 +14,13 @@ TEST_CREDENTIALS = {
 class AddWhatsAppTestCase(BaseTestCase):
     def setUp(self):
         super().setUp()
-        self.url = "/projects/%s/add_whatsapp/" % self.project.code
+        self.check = Check.objects.create(project=self.project)
+        self.url = f"/projects/{self.project.code}/add_whatsapp/"
 
     def test_instructions_work(self):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.get(self.url)
+        self.assertContains(r, "Add WhatsApp Integration")
         self.assertContains(r, "Get a WhatsApp message")
 
     @override_settings(USE_PAYMENTS=True)
@@ -50,20 +52,8 @@ class AddWhatsAppTestCase(BaseTestCase):
         self.assertTrue(c.whatsapp_notify_up)
         self.assertEqual(c.project, self.project)
 
-    def test_it_obeys_up_down_flags(self):
-        form = {"label": "My Phone", "phone": "+1234567890"}
-
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(self.url, form)
-        self.assertRedirects(r, self.channels_url)
-
-        c = Channel.objects.get()
-        self.assertEqual(c.kind, "whatsapp")
-        self.assertEqual(c.phone_number, "+1234567890")
-        self.assertEqual(c.name, "My Phone")
-        self.assertFalse(c.whatsapp_notify_down)
-        self.assertFalse(c.whatsapp_notify_up)
-        self.assertEqual(c.project, self.project)
+        # Make sure it calls assign_all_checks
+        self.assertEqual(c.checks.count(), 1)
 
     @override_settings(TWILIO_USE_WHATSAPP=False)
     def test_it_obeys_use_whatsapp_flag(self):
@@ -78,3 +68,20 @@ class AddWhatsAppTestCase(BaseTestCase):
         self.client.login(username="bob@example.org", password="password")
         r = self.client.get(self.url)
         self.assertEqual(r.status_code, 403)
+
+    def test_it_handles_down_false_up_true(self):
+        form = {"phone": "+1234567890", "up": True}
+
+        self.client.login(username="alice@example.org", password="password")
+        self.client.post(self.url, form)
+
+        c = Channel.objects.get()
+        self.assertFalse(c.whatsapp_notify_down)
+        self.assertTrue(c.whatsapp_notify_up)
+
+    def test_it_rejects_unchecked_up_and_down(self):
+        form = {"phone": "+1234567890"}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, form)
+        self.assertContains(r, "Please select at least one.")